Feat/limit client max body size

+++ Settings.hpp
@@ -34,7 +34,7 @@ typedef struct ServerBlock {
 
   int allow;
   int autoindex;
-  int client_max_body_size;
+  size_t client_max_body_size;
   std::string default_type;
   std::string index;
   int keepalive_timeout;
